done todos delete with all 3-roles

+++ src/routes/todos.js
@@ -5,7 +5,7 @@ const AccessControl = require('../models/access-control')
 
 module.exports = (db, amqpService) => {
   const router = express.Router()
-  
+  function isInteger(n) { return /^\+?(0|[1-9]\d*)$/.test(n); }
   /**
    * @openapi
    * components:
@@ -37,18 +37,15 @@ module.exports = (db, amqpService) => {
    * components:
    *  schemas:
    *    PublishEmail:
-   *      type: object
-   *      required:
-   *        - todo_id
-   *        - emails
-   *      properties:
-   *        todo_id:
-   *          type: integer
-   *        emails:
-   *          type: array
-   *          items:
+   *      type: array
+   *      items:
+   *        type: object
+   *        properties:
+   *          email:
    *            type: string
-   *            format: email 
+   *            format: email
+   *          role:
+   *            type: string
    */
 
   /**
@@ -86,11 +83,11 @@ module.exports = (db, amqpService) => {
    *              $ref: '#/components/schemas/Todo'
    */
   router.post('/', async (req, res, next) => {
-    const {uid, username } = req
-    const { title, due_date, is_completed, is_deleted} = req.body
+    const { uid, username } = req
+    const { title, due_date, is_completed, is_deleted } = req.body
     const newTodo = new Todo({ title, updated_by: username, due_date, is_completed, is_deleted })
     const todo = await db.insertTodo(newTodo)
-    const newAccessControl = new AccessControl({todo_id : todo.todo_id, user_id: uid, role: 'creator'})
+    const newAccessControl = new AccessControl({ todo_id: todo.todo_id, user_id: uid, role: 'creator' })
     await db.insertAccessControl(newAccessControl)
     res.status(201).json(todo)
   })
@@ -119,7 +116,7 @@ module.exports = (db, amqpService) => {
    *              $ref: '#/components/schemas/Error'
    */
   router.get('/', async (req, res, next) => {
-    const {uid } = req
+    const { uid } = req
     const todos = await db.findAllTodosByUid(uid)
     res.status(200).json(todos)
   })
@@ -152,19 +149,25 @@ module.exports = (db, amqpService) => {
    *              $ref: '#/components/schemas/Error'
    */
   router.get('/:id', async (req, res, next) => {
-      const { uid } = req
-      const todo_id = Number(req.params.id)
-      const authorised = await db.findAccessControlByTodoidUid(todo_id, uid)
-      if (authorised === null || authorised.role === 'read-only') {
-        res.status(403).json({error: `User not authorised to access todo_id ${todo_id}`})
+    const { uid } = req
+    const todo_id = Number(req.params.id)
+
+    if (!isInteger(todo_id)) {
+      res.status(400).json({ error: `Please provide a valid todo_id` })
+      return
+    }
+
+    const authorised = await db.findAccessControlByTodoidUid(todo_id, uid)
+    if (authorised === null || authorised.role === 'read-only') {
+      res.status(403).json({ error: `User not authorised to access todo_id ${todo_id}` })
+    } else {
+      const todo = await db.findTodoByTodoidUid(todo_id, uid);
+      if (todo) {
+        res.json(todo)
       } else {
-        const todo = await db.findTodoByTodoidUid(todo_id, uid);
-        if (todo) {
-          res.json(todo)
-        } else {
-          res.status(400).json({error: `Todo_id ${todo_id} not found`})
-        }
+        res.status(400).json({ error: `Todo_id ${todo_id} not found` })
       }
+    }
   })
 
   // router.get('/:id', async (req, res, next) => {
@@ -212,13 +215,21 @@ module.exports = (db, amqpService) => {
    *              $ref: '#/components/schemas/Error'
    */
   router.put('/:id', async (req, res, next) => {
+
+
     const { uid, username } = req
     const todo_id = Number(req.params.id)
+
+    if (!isInteger(todo_id)) {
+      res.status(400).json({ error: `Please provide a valid todo_id` })
+      return
+    }
+
     const authorised = await db.findAccessControlByTodoidUid(todo_id, uid)
     if (authorised === null || authorised.role === 'read-only') {
-      res.status(403).json({error:`User not authorised to update todo_id ${todo_id}`})
+      res.status(403).json({ error: `User not authorised to update todo_id ${todo_id}` })
     } else {
-      const { title, due_date, is_completed, is_deleted} = req.body
+      const { title, due_date, is_completed, is_deleted } = req.body
       const updatedTodo = new Todo({ todo_id, title, updated_by: username, due_date, is_completed, is_deleted })
       const todo = await db.updateTodo(todo_id, updatedTodo)
       res.status(200).json(todo)
@@ -264,21 +275,37 @@ module.exports = (db, amqpService) => {
    *              $ref: '#/components/schemas/Error'
    */
   router.delete('/:id', async (req, res, next) => {
-      const { uid } = req
-      const todo_id = Number(req.params.id)
-      const authorised = await db.findAccessControlByTodoidUid(todo_id, uid)
-      if (authorised === null || authorised.role === 'read-only') {
-        res.status(403).json({error: `User not authorised to update todo_id ${todo_id}`})
-      } else {
-        const todo = await db.deleteTodo(todo_id)
-        todo ? res.status(200).json(todo): res.status(404).json({error: `Item id ${todo_id} not found`});
+    const { uid } = req
+    const todo_id = Number(req.params.id)
+
+    if (!isInteger(todo_id)) {
+      res.status(400).json({ error: `Please provide a valid todo_id` })
+      return
+    }
+
+    const authorised = await db.findAccessControlByTodoidUid(todo_id, uid)
+    if (authorised === null || authorised.role === 'read-only') {
+      res.status(403).json({ error: `User not authorised to delete todo_id ${todo_id}` })
+    } else {
+
+      let ac = false
+      if (authorised.role === "creator") {
+        ac = await db.deleteAccessControlByTodoid(authorised.todo_id)
       }
+
+      if (authorised.role === "collaborator") {
+        ac = await db.deleteAccessControl(authorised.access_id)
+      }
+
+      const todo = await db.deleteTodo(todo_id)
+      todo ? res.status(200).json({ access_control_deleted: ac, todo }) : res.status(404).json({ error: `Item id ${todo_id} not found` });
+    }
   })
 
 
   /**
    * @openapi
-   * /todos/{id}/emails:
+   * /todos/{id}/share:
    *  post:
    *    tags:
    *    - todos
@@ -297,19 +324,53 @@ module.exports = (db, amqpService) => {
    *            schema:
    *              $ref: '#/components/schemas/PublishEmail'
    */
-  router.post('/:id/emails', async (req, res, next) => {
-    const {uid, username } = req
+  router.post('/:id/share', async (req, res, next) => {
+    const { uid, username } = req
     const todo_id = Number(req.params.id)
     const { emails } = req.body
 
-    emails.forEach(async (email) => {
-      console.log(email)
-      await amqpService.publishEmail(email, todo_id)
-    })
+    const privilegeCheck = await db.findAccessControlByTodoidUid(todo_id, uid)
+    if (!privilegeCheck) {
+      res.status(403).json({ error: `No access-privilege to todo_id ${todo_id}` })
+      return
+    }
 
-    // console.log(`todo_id: ${todo_id}, email : ${emails}`)
-    res.status(201).json({todo_id, emails})
+    //check whether emails is an Arrray
+    if (!Array.isArray(emails)) {
+      res.status(400).json({ error: "Please provide emails array." })
+      return
+    }
+
+    emails.forEach(async (item) => {
+      await amqpService.publishEmail(item.email, item.role, todo_id)
+    })
+    res.status(204).json({})
   })
+
+
+
+  // router.post('/:id/emails', async (req, res, next) => {
+  //   const {uid, username } = req
+  //   const todo_id = Number(req.params.id)
+  //   const { emails } = req.body
+
+  //   const privilegeCheck = await db.findAccessControlByTodoidUid(todo_id, uid)
+  //   if (!privilegeCheck){
+  //     res.status(403).json({error: `No access-privilege to todo_id ${todo_id}`})
+  //     return
+  //   } 
+
+  //   //check whether emails is an Arrray
+  //   if (!Array.isArray(emails)){
+  //     res.status(400).json({error: "Please provide emails array."})
+  //     return
+  //   }
+
+  //   emails.forEach(async (email) => {
+  //     await amqpService.publishEmail(email,"collaborator", todo_id)
+  //   })
+  //   res.status(204).json({})
+  // })
 
 
   return router
